+++ backend/config/config.exs
@@ -68,8 +68,7 @@ config :backend, :crawler,
   user_agent: "fediverse.space crawler",
   admin_phone: System.get_env("ADMIN_PHONE"),
   twilio_phone: System.get_env("TWILIO_PHONE"),
-  admin_email: System.get_env("ADMIN_EMAIL"),
-  frontend_domain: "https://www.fediverse.space"
+  admin_email: System.get_env("ADMIN_EMAIL")
 
 config :backend, Backend.Scheduler,
   jobs: [