+++ CHANGELOG.md
@@ -9,17 +9,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Clarify that the admin page only works for Mastodon and Pleroma instances.
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [2.3.1 - 2019-08-03]
+
+### Added
+
 - Added a warning on mobile devices suggesting to view the site on a larger computer.
 
 ### Changed
 
 - Performance improvements when opening the app on something that isn't the graph.
 - There are now fewer irrelevant search results.
-
-### Deprecated
-
-### Removed
+- Clarify that the admin page only works for Mastodon and Pleroma instances.
 
 ### Fixed